+++ clients/da-vinci-client/src/main/java/com/linkedin/davinci/blobtransfer/P2PBlobTransferManager.java
@@ -1,4 +1,4 @@
-package com.linkedin.venice.blobtransfer;
+package com.linkedin.davinci.blobtransfer;
 
 import com.linkedin.venice.exceptions.VeniceException;
 import java.util.concurrent.CompletionStage;
